Move tests for conversion into pandas into separate file
