+++ docs/api/frame/to_pandas.rst
@@ -2,3 +2,4 @@
 .. xmethod:: datatable.Frame.to_pandas
     :src: src/core/frame/to_numpy.cc Frame::to_pandas
     :doc: src/core/frame/to_numpy.cc doc_to_pandas
+    :tests: tests/frame/test-to-pandas.py
